+++ app/assemblers/assemble_reputation_tokens.rb
@@ -1,0 +1,44 @@
+class AssembleReputationTokens
+  include Mandate
+
+  initialize_with :user, :params
+
+  def self.keys
+    %w[criteria category order per_page page]
+  end
+
+  def call
+    data = tokens.map do |token|
+      token.rendering_data.merge(
+        links: {
+          mark_as_seen: Exercism::Routes.mark_as_seen_api_reputation_url(token.uuid)
+        }
+      )
+    end
+
+    SerializePaginatedCollection.(
+      tokens,
+      data: data,
+      meta: {
+        links: {
+          tokens: Exercism::Routes.reputation_journey_url,
+          mark_all_as_seen: Exercism::Routes.mark_all_as_seen_api_reputation_index_url
+        },
+        total_reputation: user.formatted_reputation,
+        is_all_seen: user.reputation_tokens.unseen.empty?
+      }
+    )
+  end
+
+  memoize
+  def tokens
+    User::ReputationToken::Search.(
+      user,
+      criteria: params[:criteria],
+      category: params[:category],
+      order: params[:order],
+      per: params[:per_page],
+      page: params[:page]
+    )
+  end
+end
